Add 'vm_up' metric for each VM in prometheus metrics
The new 'vm_up' is there for each VM and reports 1 when the VM is in the
state "Running". If the VM is in any other state or is shut off it will
be report 0.

Note that a VM will be reported as down (0) for 10 minutes if it just
disappears or is shut off, to give prometheus the possibility to catch
up with that. Otherwise prometheus would report the VM up for another 5
minutes if the scrape target is still reachable.